tests/wpt/mozilla/tests/bluetooth/getCharacteristics/device-goes-out-of-range.https.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/_mozilla/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(t => {
    window.testRunner.setBluetoothMockDataSet(adapter_type.heart_rate);
    return window.navigator.bluetooth.requestDevice({
        filters: [{services: [heart_rate.name]}]
    })
    .then(device => device.gatt.connect())
    .then(gattServer => gattServer.getPrimaryService(heart_rate.name))
    .then(service => {
        window.testRunner.setBluetoothMockDataSet(adapter_type.empty);
        return promise_rejects_dom(t, 'NetworkError', service.getCharacteristics());
    });
}, 'Device goes out of range. Reject with NetworkError.');
</script>
